+++ Changes.md
@@ -6,7 +6,7 @@
 
 *   Source upgrades for ForgeGradle 3.x and Gradle on 1.12.2 and above
 *   Added a `Format Words` option in Advanced Settings, used to toggle automatic capitalization in RPC Strings
-*   Added `Entity Target` and `Entity Riding` Options in Advanced Settings to begin the Per-Entity System
+*   Added `Entity Target`, `Entity Attacking` and `Entity Riding` Options in Advanced Settings to begin the Per-Entity System
 *   Changed `&ENTITY&` placeholder belonging to Items and Blocks to `&TILEENTITY&`
 *   Backend Upgrades and Rewrites to Gui Screens and related Utilities
 *   Localization Updates to Update Checker